Make walkthrough test after() async/await (#27)
Add async/await clauses to the test.js file within the firestore walkthrough.

I took this example and extended it to match my project and use case... it ended up running into timing collisions and flaky tests without this  async/await in the after hooks.  Therefore, it made sense to update the source to help the next person.